ushort に修正する

+++ Assets/ArtNet/Runtime/Scripts/Devices/IDmxDevice.cs
@@ -3,8 +3,8 @@ namespace ArtNet.Devices
     public interface IDmxDevice
     {
         byte ChannelNumber { get; }
-        int Universe { get; }
-        int StartAddress { get; }
+        ushort Universe { get; }
+        ushort StartAddress { get; }
         void DmxUpdate(byte[] dmx);
     }
 }
